requirements.txt
PyQt6>=6.6.1                    
openpyxl==3.1.2               
reportlab==4.0.7               
argon2-cffi==23.1.0           
python-dateutil==2.8.2        
